Feat: Made nix-package script posix sh

+++ software/scripts/nix-package.sh
@@ -1,14 +1,15 @@
-#!/usr/bin/env bash
+#!/bin/sh
 
-# If you're changing this file, make sure to read the systems/software/scripts docs
+TOP_LEVEL="$(git rev-parse --show-toplevel || exit)"
 
-set -euo pipefail
+# cd to the nix directory
+if ! cd "$TOP_LEVEL/nix"; then
+  echo "Error when entering nix directory"
+  exit
+fi
 
-# cd to the software directory
-cd "$(git rev-parse --show-toplevel)"/software
-
-ROS_WS="ros_ws"
-OUTPUT_DIR="$ROS_WS"/nix-packages
+ROS_WS="$TOP_LEVEL/software/ros_ws"
+OUTPUT_DIR="$TOP_LEVEL/nix/ros-packages/perseus"
 
 echo "Cleaning old packaging files"
 rm -r "$OUTPUT_DIR"
@@ -21,32 +22,42 @@ echo "Running ros2nix to generate new packaging files..."
 # nix run is the "better" way to do this, but it doesn't allow for complete environment isolation like the shell -i flag.
 # we want word splitting here to get all the package files in one go, so disable the shellcheck warning
 # shellcheck disable=SC2046
-nix shell -i github:wentasah/ros2nix --command ros2nix --output-dir="$OUTPUT_DIR" --output-as-nix-pkg-name --no-default --distro jazzy $(find "$ROS_WS" -type d \( -name log -o -name build -o -name install \) -prune -o -name 'package.xml' -print)
-echo "Formatting generated files"
-cd "$OUTPUT_DIR"
-git add .
-nix fmt
+if ! nix shell -i github:wentasah/ros2nix --command ros2nix --output-dir="$OUTPUT_DIR" --output-as-nix-pkg-name --no-default --distro jazzy $(find "$ROS_WS" -type d \( -name log -o -name build -o -name install \) -prune -o -name 'package.xml' -print); then
+  echo "Error when generating nix derivations"
+  exit
+fi
 
-if [[ $* == *--no-commit* ]]; then
+echo "Formatting generated files"
+if ! cd "$OUTPUT_DIR"; then
+  echo "Error when entering generated packages output directory"
+  exit
+fi
+git add .
+treefmt "$OUTPUT_DIR"
+
+if [ "$*" = "*--no-commit*" ]; then
   echo "Will not commit changes"
   exit 0
 fi
 
-if ! git diff --cached --quiet >/dev/null; then
-  HAS_GIT_STAGING=1
-  echo "Stashing staged changes"
-  git stash push -S >/dev/null
-fi
-
 cleanup() {
-  if [[ -v HAS_GIT_STAGING ]]; then
+  if [ -n "$HAS_GIT_STAGING" ]; then
     echo "Restoring git stash"
     git stash pop >/dev/null
   fi
 }
+
 trap cleanup EXIT
+
+# Pathspec: https://git-scm.com/docs/gitglossary#Documentation/gitglossary.txt-aiddefpathspecapathspec
+# TODO: Test this
+if ! git diff --cached --quiet -- "$TOP_LEVEL" ":!$OUTPUT_DIR" >/dev/null; then
+  HAS_GIT_STAGING=1
+  echo "Stashing staged changes"
+  git stash push -S -- "$TOP_LEVEL" ":!$OUTPUT_DIR" >/dev/null
+fi
 
 echo "Staging changes"
 git add .
 echo "Committing changes"
-git commit -m "chore: Update Nix packaging" >/dev/null
+git commit -m "Chore: Update Nix packaging" >/dev/null
